Give more output via syslog for various cases (#455346)
yum-updatesd should log the number of updates available (which it does); the
number of updates applied (if 'do_update = yes' is set); and any error messages
if the updates fail.

In our case we configure /etc/yum.conf on servers to exclude some packages.
However, if these cause a dependency failure for other updates, then
yum-updatesd should log this. The patch corrects this.

The problem we found was that several servers had updates to apply, but it
seemed that yum-updatesd was not applying them. Nothing was logged in
/var/log/messages at all. It turned out that there was a dependency failure.
Applying our patch, then yum-updatesd logged the failure.